refactor(epicenter): invert control for createServer and createCLI
- createServer now takes client, returns { app, start } instead of creating client internally
- createCLI now takes client, returns { run } instead of runCLI({ client, argv })
- Delete cli/server.ts, merge functionality into server.ts
- Update all tests to use new API patterns
- Skip MCP tests (pre-existing: endpoint disabled)
- Fix test setup for Elysia server port access
