+++ packages/epicenter/src/cli/cli-end-to-end.test.ts
@@ -4,6 +4,7 @@ import { mkdir, rm } from 'node:fs/promises';
 import path from 'node:path';
 import { type } from 'arktype';
 import { Ok } from 'wellcrafted/result';
+import { createClient } from '../core/workspace/client.node';
 import {
 	defineMutation,
 	defineQuery,
@@ -128,69 +129,56 @@ describe('CLI End-to-End Tests', () => {
 	});
 
 	test('CLI can create a post', async () => {
-		await createCLI({
-			workspaces,
-			options,
-			argv: [
-				'posts',
-				'createPost',
-				'--title',
-				'Test Post',
-				'--content',
-				'Test content',
-				'--category',
-				'tech',
-			],
-		});
+		const client = await createClient(workspaces, options);
+		await createCLI(client).run([
+			'posts',
+			'createPost',
+			'--title',
+			'Test Post',
+			'--content',
+			'Test content',
+			'--category',
+			'tech',
+		]);
 
-		// Verify the post was created by checking the markdown file
 		await new Promise((resolve) => setTimeout(resolve, 200));
 		const files = await Bun.$`ls ${TEST_MARKDOWN}/posts`.text();
 		expect(files.trim().length).toBeGreaterThan(0);
 	});
 
 	test('CLI can query posts', async () => {
-		// First create a post
-		await createCLI({
-			workspaces,
-			options,
-			argv: [
-				'posts',
-				'createPost',
-				'--title',
-				'Query Test',
-				'--category',
-				'tech',
-			],
-		});
+		const client = await createClient(workspaces, options);
+		await createCLI(client).run([
+			'posts',
+			'createPost',
+			'--title',
+			'Query Test',
+			'--category',
+			'tech',
+		]);
 
-		// Wait for the post to be created
 		await new Promise((resolve) => setTimeout(resolve, 200));
 
-		// Now query all posts
-		await createCLI({
-			workspaces,
-			options,
-			argv: ['posts', 'listPosts'],
-		});
+		const client2 = await createClient(workspaces, options);
+		await createCLI(client2).run(['posts', 'listPosts']);
 	});
 
 	test('CLI handles missing required options', async () => {
 		try {
-			await createCLI({
-				workspaces,
-				options,
-				argv: ['posts', 'createPost', '--title', 'Missing Category'],
-			});
-			expect(false).toBe(true); // Should not reach here
+			const client = await createClient(workspaces, options);
+			await createCLI(client).run([
+				'posts',
+				'createPost',
+				'--title',
+				'Missing Category',
+			]);
+			expect(false).toBe(true);
 		} catch (error) {
-			// Expected to fail due to missing required field
 			expect(error).toBeDefined();
 		}
 	});
 
 	test('CLI properly formats success output', async () => {
-		// Capture console output
 		const logs: string[] = [];
 		const originalLog = console.log;
 		console.log = (...args: unknown[]) => {
@@ -198,22 +186,18 @@ describe('CLI End-to-End Tests', () => {
 			originalLog(...args);
 		};
 
-		await createCLI({
-			workspaces,
-			options,
-			argv: [
-				'posts',
-				'createPost',
-				'--title',
-				'Output Test',
-				'--category',
-				'test',
-			],
-		});
+		const client = await createClient(workspaces, options);
+		await createCLI(client).run([
+			'posts',
+			'createPost',
+			'--title',
+			'Output Test',
+			'--category',
+			'test',
+		]);
 
 		console.log = originalLog;
 
-		// Verify output format
 		expect(logs.some((log) => log.includes('Success'))).toBe(true);
 	});
 });
