+++ packages/epicenter/src/cli/cli.ts
@@ -1,65 +1,32 @@
 import type { TaggedError } from 'wellcrafted/error';
 import { isResult, type Result } from 'wellcrafted/result';
 import yargs from 'yargs';
-import { walkActions } from '../core/actions';
-import type { AnyWorkspaceConfig } from '../core/workspace';
-import {
-	createClient,
-	type CreateClientOptions,
-	type WorkspaceClient,
-} from '../core/workspace/client.node';
-import { DEFAULT_PORT, startServer } from './server';
+import { walkActions, type WorkspaceExports } from '../core/actions';
+import type { AnyWorkspaceConfig, EpicenterClient } from '../core/workspace';
+import type { WorkspaceClient } from '../core/workspace/client.node';
+import { createServer, DEFAULT_PORT } from '../server/server';
 import { standardJsonSchemaToYargs } from './standard-json-schema-to-yargs';
 
-export type CreateCLIOptions = CreateClientOptions;
-
 /**
- * Create and run CLI from workspace configurations.
+ * Create a CLI from an initialized Epicenter client.
  *
- * This function:
- * 1. Initializes workspaces (with persistence, sync providers)
- * 2. Generates yargs command hierarchy (workspace → action)
- * 3. Parses arguments and executes the matched command
- * 4. Cleans up workspaces after command execution (including on Ctrl+C)
- *
- * The client lifecycle is managed internally to ensure persistence providers
- * remain active throughout command execution.
- *
- * @param workspaces - Array of workspace configurations
- * @param argv - Array of command-line arguments to parse
- * @param options - Optional client options (e.g., storageDir)
+ * @param client - Initialized Epicenter client from createClient()
+ * @returns Object with run method to execute CLI
  *
  * @example
  * ```typescript
- * // In production (bin.ts)
+ * import { createClient } from '@epicenter/hq';
+ * import { createCLI } from '@epicenter/hq/cli';
  * import { hideBin } from 'yargs/helpers';
- * await createCLI({ workspaces, argv: hideBin(process.argv) });
  *
- * // In tests
- * await createCLI({ workspaces, argv: ['posts', 'createPost', '--title', 'Test'] });
+ * const client = await createClient(workspaces, options);
+ * await createCLI(client).run(hideBin(process.argv));
  * ```
  */
-export async function createCLI<
+export function createCLI<
 	const TWorkspaces extends readonly AnyWorkspaceConfig[],
->({
-	workspaces,
-	argv,
-	options,
-}: {
-	workspaces: TWorkspaces;
-	argv: string[];
-	options?: CreateCLIOptions;
-}): Promise<void> {
-	const client = await createClient(workspaces, options);
-
-	const cleanup = async () => {
-		await client.destroy();
-		process.exit(0);
-	};
-	process.on('SIGINT', cleanup);
-	process.on('SIGTERM', cleanup);
-
-	let cli = yargs(argv)
+>(client: EpicenterClient<TWorkspaces>) {
+	let cli = yargs()
 		.scriptName('epicenter')
 		.usage('Usage: $0 [command] [options]')
 		.help()
@@ -76,23 +43,27 @@ export async function createCLI<
 				default: DEFAULT_PORT,
 			});
 		},
-		async (argv) => {
-			await startServer(workspaces, { ...options, port: argv.port });
+		(argv) => {
+			createServer(client).start({ port: argv.port });
 		},
 	);
 
-	for (const workspaceConfig of workspaces) {
-		const workspaceId = workspaceConfig.id;
-		// Cast to indexed access since we know workspaceId exists (client was created from workspaces)
-		const workspaceClient = (client as Record<string, WorkspaceClient<any>>)[
-			workspaceId
-		]!;
+	const {
+		destroy: _clientDestroy,
+		[Symbol.asyncDispose]: _clientDispose,
+		...workspaceClients
+	} = client;
 
+	for (const [workspaceId, workspaceClient] of Object.entries(
+		workspaceClients,
+	)) {
+		const typedClient = workspaceClient as WorkspaceClient<WorkspaceExports>;
 		const {
 			destroy: _,
 			[Symbol.asyncDispose]: __,
+			$ydoc: ___,
 			...workspaceExports
-		} = workspaceClient;
+		} = typedClient;
 
 		cli = cli.command(
 			workspaceId,
@@ -148,11 +119,22 @@ export async function createCLI<
 		);
 	}
 
-	try {
-		await cli.parse();
-	} finally {
-		process.off('SIGINT', cleanup);
-		process.off('SIGTERM', cleanup);
-		await client.destroy();
-	}
+	return {
+		async run(argv: string[]) {
+			const cleanup = async () => {
+				await client.destroy();
+				process.exit(0);
+			};
+			process.on('SIGINT', cleanup);
+			process.on('SIGTERM', cleanup);
+
+			try {
+				await cli.parse(argv);
+			} finally {
+				process.off('SIGINT', cleanup);
+				process.off('SIGTERM', cleanup);
+				await client.destroy();
+			}
+		},
+	};
 }
