+++ packages/epicenter/src/cli/index.ts
@@ -1,6 +1,3 @@
-// Programmatic CLI generation
 export { createCLI } from './cli';
 export { loadEpicenterConfig } from './load-config';
-// Server
-export { type ServeOptions, startServer } from './server';
 export { standardSchemaToYargs } from './standardschema-to-yargs';
